Remove d2l-my-courses-content-behavior
