+++ src/d2l-my-courses-container.js
@@ -98,7 +98,6 @@ class MyCoursesContainer extends mixinBehaviors([
 							<d2l-my-courses-content
 								presentation-url="[[_presentationUrl]]"
 								token="[[token]]"
-								disable-entity-cache
 								advanced-search-url="[[advancedSearchUrl]]"
 								course-image-upload-cb="[[courseImageUploadCb]]"
 								enrollments-search-action="[[item.enrollmentsSearchAction]]"
@@ -119,7 +118,6 @@ class MyCoursesContainer extends mixinBehaviors([
 				<d2l-my-courses-content
 					presentation-url="[[_presentationUrl]]"
 					token="[[token]]"
-					disable-entity-cache
 					advanced-search-url="[[advancedSearchUrl]]"
 					org-unit-type-ids="[[orgUnitTypeIds]]"
 					course-image-upload-cb="[[courseImageUploadCb]]"
